changes to Python decorator API for extensibility  as well as benchmark plugins
